+++ Changes.md
@@ -4,6 +4,10 @@ Changes
 V405.1.7
 ---------
 1. End of week 8.
+2. Improve AMD JS, refs:
+   - https://moodledev.io/docs/4.5/guides/javascript#working-with-strings
+   - https://docs.moodle.org/dev/Javascript/Coding_Style
+   - admin/tool/dataprivacy/amd/src/categoriesactions.js
 
 V405.1.6
 ---------